requirements.txt
pandas
python-dotenv
Pillow
requests
AsciiArt

